#1204 Ruff Formatter (#1582)

+++ app/dao/fido2_key_dao.py
@@ -2,9 +2,7 @@ from app import db
 from app.models import Fido2Key, Fido2Session
 from app.config import Config
 
-from app.dao.dao_utils import (
-    transactional
-)
+from app.dao.dao_utils import transactional
 
 from sqlalchemy import and_
 
@@ -14,23 +12,23 @@ import pickle  # nosec
 import base64
 
 
-def delete_fido2_key(user_id, id):
-    db.session.query(Fido2Key).filter(
-        and_(Fido2Key.user_id == user_id, Fido2Key.id == id)
-    ).delete()
+def delete_fido2_key(
+    user_id,
+    id,
+):
+    db.session.query(Fido2Key).filter(and_(Fido2Key.user_id == user_id, Fido2Key.id == id)).delete()
     db.session.commit()
 
 
-def get_fido2_key(user_id, id):
-    return Fido2Key.query.filter(
-        and_(Fido2Key.user_id == user_id, Fido2Key.id == id)
-    ).one()
+def get_fido2_key(
+    user_id,
+    id,
+):
+    return Fido2Key.query.filter(and_(Fido2Key.user_id == user_id, Fido2Key.id == id)).one()
 
 
 def list_fido2_keys(user_id):
-    return Fido2Key.query.filter(
-        Fido2Key.user_id == user_id
-    ).order_by(Fido2Key.created_at.asc()).all()
+    return Fido2Key.query.filter(Fido2Key.user_id == user_id).order_by(Fido2Key.created_at.asc()).all()
 
 
 @transactional
@@ -39,35 +37,31 @@ def save_fido2_key(fido2_key):
 
 
 @transactional
-def create_fido2_session(user_id, session):
+def create_fido2_session(
+    user_id,
+    session,
+):
     delete_fido2_session(user_id)
-    db.session.add(
-        Fido2Session(user_id=user_id, session=json.dumps(session))
-    )
+    db.session.add(Fido2Session(user_id=user_id, session=json.dumps(session)))
 
 
 def delete_fido2_session(user_id):
-    db.session.query(Fido2Session).filter(
-        Fido2Session.user_id == user_id
-    ).delete()
+    db.session.query(Fido2Session).filter(Fido2Session.user_id == user_id).delete()
 
 
 def get_fido2_session(user_id):
-    session = db.session.query(Fido2Session).filter(
-        Fido2Session.user_id == user_id
-    ).one()
+    session = db.session.query(Fido2Session).filter(Fido2Session.user_id == user_id).one()
     delete_fido2_session(user_id)
     return json.loads(session.session)
 
 
-def decode_and_register(data, state):
+def decode_and_register(
+    data,
+    state,
+):
     client_data = CollectedClientData(data['clientDataJSON'])
     att_obj = AttestationObject(data['attestationObject'])
 
-    auth_data = Config.FIDO2_SERVER.register_complete(
-        state,
-        client_data,
-        att_obj
-    )
+    auth_data = Config.FIDO2_SERVER.register_complete(state, client_data, att_obj)
 
     return base64.b64encode(pickle.dumps(auth_data.credential_data)).decode('utf8')
